requirements.txt
anyio==4.3.0
beautifulsoup4==4.12.3
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
telegram_text==0.2.0
typing_extensions==4.12.0
urllib3==2.2.1
